Refactor leverage slider and seperate calculator
